Сборка скриптов в один файл
