+++ package.json
@@ -9,6 +9,7 @@
     "browser-sync": "2.26.x",
     "del": "^4.1.0",
     "gulp": "4.0.x",
+    "gulp-concat": "^2.6.1",
     "gulp-csso": "^3.0.1",
     "gulp-imagemin": "^5.0.3",
     "gulp-plumber": "1.2.x",
