+++ source/catalog.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link  href="css/style.min.css" rel="stylesheet">
     <title>«Мишка» - каталог товаров</title>
-    <script src="js/lib/picturefill.min.js" async></script>
   </head>
   <body>
     <div style="display: none;">
@@ -341,8 +340,7 @@
       </section>
     </div>
 
-    <script src="js/app/fix-ie-min-height.min.js"></script>
-    <script src="js/app/mobile-menu.min.js"></script>
-    <script src="js/app/add-to-basket.min.js"></script>
+    <script src="js/picturefill.min.js" async></script>
+    <script src="js/app.min.js"></script>
   </body>
 </html>
